Update documentParser.ts
Hi, I can't test it because I'm not comfortable with typescript (should you migrate it to js? 👍🏻 ). This array doesn't have waveforms.

sequenceSet[] has 13 components inside and you need to exclude the first and consider the other 12.

I think this is the cause to don't be able to plot on your website.